requirements.txt
asgiref==3.8.1
boto3==1.35.76
botocore==1.35.76
Django==4.2
djangorestframework==3.15.2
environs==11.2.1
jmespath==1.0.1
marshmallow==3.23.1
packaging==24.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==1.26.14
gunicorn==23.0.0